@ikawrakow 2025-05-14
> Well, that's amazing news, even if your sponsor doesn't get back to you.

Haha, this is because you don't know me, and so don't expect for how long I'm going to procrastinate on this.

> Do you think TP is achievable here?

What is TP?