@ikawrakow 2025-05-14
> Without speculative decoding, 2x3090@275w:
> 
>     * Llama 3.3 70B 4.5bpw, from 18.1 to 22.9 tok/s
> 
>     * Mistral Large 123B 3.0bpw, from 15.5 to 22.3 tok/s
> 
> 
> Exl3 is supposed to have even better TP performance, but it's not implemented yet.

So, barely faster than `llama.cpp`? I have a 4080 (717 GB/s), so less bandwidth than a 3090 (935 GB/s), and I get 125 t/s for Llama-8B at 4.5 bpw on the 4080. Napkin math: `125 * 8/70 * 935/717 = 18.6 t/s`.