@ThomasBaruzier 2025-05-28
Thanks for all the feedback!

FA helps with 4+bpw as you predicted, but for i- and iqk-quants, I'll investigate further another time, maybe a few param tweaks could help?

Here is a refined plot:
![ploty](https://github.com/user-attachments/assets/14a96d32-cc1b-460f-bd92-93b258f61af5)